Rename "dep.ts" to "deps.ts"
